% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dejong-map.R
\name{dejong_map}
\alias{dejong_map}
\title{The De Jong map.}
\usage{
dejong_map(n = 1000, xi = -5, yi = 0.1, a = 1.4, b = -2.3, c = 2.4,
  d = -2.1, add.timestep = FALSE, plot = FALSE)
}
\arguments{
\item{n}{Number of points in the time series.}

\item{xi}{The initial value for x}

\item{yi}{The initival value for y}

\item{a}{The value of the parameter a}

\item{b}{The value of the parameter b}

\item{c}{The value of the parameter c}

\item{d}{The value of the parameter d}

\item{add.timestep}{Add a column indicating the time step?}

\item{plot}{Plot the time series?}
}
\value{
A n-by-2 data frame containing the values of the x and y components.
}
\description{
The De Jong map.
}
