% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gingerbreadman-map.R
\name{gingerbreadman_map}
\alias{gingerbreadman_map}
\title{Gingerbreadman map.}
\usage{
gingerbreadman_map(n = 1000, xi = -0.2, yi = 0.18, add.timestep = FALSE,
  plot = FALSE)
}
\arguments{
\item{n}{Number of points in the time series.}

\item{xi}{The initial value for x}

\item{yi}{The initival value for y}

\item{add.timestep}{Add a column indicating the time step?}

\item{plot}{Plot the time series?}
}
\value{
A dataframe containing columns time series for x and y
    (and optionally "t", a timestep column).

A n-by-2 data frame containing the values of the x and y components.
}
\description{
https://en.wikipedia.org/wiki/Gingerbreadman_map
Devaney, Robert L. (1988), "Fractal patterns arising in chaotic dynamical systems",
in Peitgen, Heinz-Otto; Saupe, Dietmar, The Science of Fractal Images, Springer-Verlag,
pp. 137–168, doi:10.1007/978-1-4612-3784-6_3. See in particular Fig. 3.3.
}
