% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/henon-map.R
\name{henon_map}
\alias{henon_map}
\title{The Henon map.}
\usage{
henon_map(n = 1000, xi = 1, yi = 1, a = 1.4, b = 0.3,
  add.timestep = FALSE, plot = FALSE)
}
\arguments{
\item{n}{Number of points in the time series.}

\item{xi}{The initial value for x}

\item{yi}{The initival value for y}

\item{a}{The value of the parameter a}

\item{b}{The value of the parameter b}

\item{add.timestep}{Add a column indicating the time step?}

\item{plot}{Plot the time series?}
}
\value{
A n-by-2 data frame containing the values of the x and y components.
}
\description{
The Henon map.
}
