% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ikeda-map.R
\name{ikeda_map}
\alias{ikeda_map}
\title{Ikeda map.}
\usage{
ikeda_map(n = 1000, xi = -5, yi = 0.1, u = 0.6, add.timestep = FALSE,
  plot = FALSE)
}
\arguments{
\item{n}{Number of points in the time series.}

\item{xi}{The initial value for x}

\item{yi}{The initival value for y}

\item{u}{The value of the parameter u}

\item{add.timestep}{Add a column indicating the time step?}

\item{plot}{Plot the time series?}
}
\value{
A n-by-2 data frame containing the values of the x and y components.
}
\description{
K.Ikeda, Multiple-valued Stationary State and its Instability of the Transmitted Light by a Ring Cavity System, Opt. Commun. 30 257-261 (1979)
}
\details{
The Ikeda map. From Ikeda, H. Daido and O. Akimoto, Optical Turbulence:
Chaotic Behavior of Transmitted Light from a Ring Cavity,
Phys. Rev. Lett. 45, 709–712 (1980).
}
