% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaplanyorke-map.R
\name{kaplanyorke_map}
\alias{kaplanyorke_map}
\title{The tinkerbell map. From the Wikipedia article on chaotic maps.}
\usage{
kaplanyorke_map(n = 1000, xi = 0.5, yi = 0.4, b = 104123,
  alpha = 0.01, add.timestep = FALSE, plot = FALSE)
}
\arguments{
\item{n}{Number of points in the time series.}

\item{xi}{The initial value for x}

\item{yi}{The initival value for y}

\item{b}{A large prime number}

\item{add.timestep}{Add a column indicating the time step?}

\item{plot}{Plot the time series?}

\item{a}{The value of the parameter a}
}
\value{
A n-by-2 data frame containing the values of the x and y components.
}
\description{
The tinkerbell map. From the Wikipedia article on chaotic maps.
}
