% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic-map-five-species-sugihara.R
\name{logistic_map_fivespecies_sugihara}
\alias{logistic_map_fivespecies_sugihara}
\title{Time series for a two-species model system. Unidirectional forcing from x introduces synchrony in y.}
\usage{
logistic_map_fivespecies_sugihara(n = 3000, y1.i = 0.4, y2.i = 0.4,
  y3.i = 0.4, y4.i = 0.4, y5.i = 0.4, Ry1y1 = 4, Ry1y2 = 0.31,
  Ry1y3 = 0.636, Ry1y4 = 0.111, Ry1y5 = 0.082, Ry2y2 = 3.1, Ry2y1 = 2,
  Ry2y3 = 0.636, Ry2y4 = 0.011, Ry2y5 = 0.111, Ry3y3 = 2.12,
  Ry3y1 = 0.4, Ry3y2 = 0.93, Ry3y4 = 0.131, Ry3y5 = 0.125,
  Ry4y4 = 3.8, Ry4y5 = 4.1, Ry5y5 = 4.1, add.timestep = FALSE,
  plot = FALSE)
}
\arguments{
\item{n}{The number of time steps that will be generated.}

\item{Ry1y1}{Parameter controlling the influence of y1 at timestep i on itself at timestep i+1.}

\item{Ry1y2}{Parameter controlling the influence of y1 at timestep i on on y2 at timestep i+1.}

\item{Ry1y3}{Parameter controlling the influence of y1 at timestep i on on y3 at timestep i+1.}

\item{Ry1y4}{Parameter controlling the influence of y1 at timestep i on on y4 at timestep i+1.}

\item{Ry1y5}{Parameter controlling the influence of y1 at timestep i on on y5 at timestep i+1.}

\item{Ry2y2}{Parameter controlling the influence of y2 at timestep i on itself at timestep i+1.}

\item{Ry2y1}{Parameter controlling the influence of y2 at timestep i on on y1 at timestep i+1.}

\item{Ry2y3}{Parameter controlling the influence of y2 at timestep i on on y3 at timestep i+1.}

\item{Ry2y4}{Parameter controlling the influence of y2 at timestep i on on y4 at timestep i+1.}

\item{Ry2y5}{Parameter controlling the influence of y2 at timestep i on on y5 at timestep i+1.}

\item{Ry3y3}{Parameter controlling the influence of y3 at timestep i on itself at timestep i+1.}

\item{Ry3y1}{Parameter controlling the influence of y3 at timestep i on on y1 at timestep i+1.}

\item{Ry3y2}{Parameter controlling the influence of y3 at timestep i on on y2 at timestep i+1.}

\item{Ry3y4}{Parameter controlling the influence of y3 at timestep i on on y4 at timestep i+1.}

\item{Ry3y5}{Parameter controlling the influence of y3 at timestep i on on y5 at timestep i+1.}

\item{Ry4y4}{Parameter controlling the influence of y4 at timestep i on itself at timestep i+1.}

\item{Ry4y5}{Parameter controlling the influence of y4 at timestep i on on y5 at timestep i+1.}

\item{Ry5y5}{Parameter controlling the influence of y5 at timestep i on itself at timestep i+1.}

\item{y1}{Initial value of time series y1.}

\item{y2}{Initial value of time series y2.}

\item{y3}{Initial value of time series y3.}

\item{y4}{Initial value of time series y4.}

\item{y5}{Initial value of time series y5.}

\item{Ry4y1}{Parameter controlling the influence of y4 at timestep i on on y1 at timestep i+1.}

\item{Ry4y2}{Parameter controlling the influence of y4 at timestep i on on y2 at timestep i+1.}

\item{Ry4y3}{Parameter controlling the influence of y4 at timestep i on on y3 at timestep i+1.}
}
\value{
A dataframe containing columns time series for x and y
    (and optionally "t", a timestep column).

Subsystem y1, y2 and y3 is the forcing subsystem, that interacts unidirectionally
with y4 and y5. y4 and y5 do not interact with each other and do not influence y1, y2 and
y3. Initial conditions are not provided in the original paper, but are all set to 0.4 here.
}
\description{
Source: Sugihara, G., May, R., Ye, H., Hsieh, C. H., Deyle, E., Fogarty, M., & Munch, S. (2012).
Detecting causality in complex ecosystems. science, 338(6106), 496-500.
}
