% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic-map-fourspecies-transitive.R
\name{logistic_map_fourspecies_transitive}
\alias{logistic_map_fourspecies_transitive}
\title{Four-species logistic map system.
Transitive unidirectional forcing (y1 -> y2 -> y3 -> y4).}
\usage{
logistic_map_fourspecies_transitive(n = 3000, y1.i = 0.4, y2.i = 0.4,
  y3.i = 0.4, y4.i = 0.4, Ry1y1 = 3.9, Ry1y2 = 0.4, Ry2y2 = 3.6,
  Ry2y3 = 0.4, Ry3y3 = 3.6, Ry3y4 = 0.35, Ry4y4 = 3.8,
  add.timestep = FALSE, plot = FALSE)
}
\arguments{
\item{n}{The number of time steps that will be generated.}

\item{y1.i}{Initial value of time series y1.}

\item{y2.i}{Initial value of time series y2.}

\item{y3.i}{Initial value of time series y3.}

\item{y4.i}{Initial value of time series y4.}

\item{Ry1y1}{Parameter controlling the influence of y1 at timestep i on itself at timestep i+1}

\item{Ry1y2}{Parameter controlling the influence of y1 at timestep i on y2 at timestep i+1}

\item{Ry2y2}{Parameter controlling the influence of y2 at timestep i on itself at timestep i+1}

\item{Ry2y3}{Parameter controlling the influence of y2 at timestep i on y3 at timestep i+1}

\item{Ry3y3}{Parameter controlling the influence of y1 at timestep i on itself at timestep i+1}

\item{Ry3y4}{Parameter controlling the influence of y3 at timestep i on y4 at timestep i+1}

\item{Ry4y4}{Parameter controlling the influence of y1 at timestep i on itself at timestep i+1}
}
\value{
A dataframe containing columns time series for y1, y2, y3 and y4
    (and optionally "t", a timestep column).

Defaults values are as in the original paper.
}
\description{
Source: Ye, H. et al. (2015). Distinguishing time-delayed causal interactions
using convergent cross mapping. Sci. Rep. 5, 14750; doi: 10.1038/srep14750 (2015)
}
\details{
This function allows you to express your love of cats.
}
