% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic-unidirectional-sugihara.R
\name{logistic_map_twospecies_unidir}
\alias{logistic_map_twospecies_unidir}
\title{Time series for a two-species (logistic) model system. Unidirectional forcing from x to y.}
\usage{
logistic_map_twospecies_unidir(n = 1000, xi = 0.2, yi = 0.4, Rx = 3.7,
  Ry = 3.7, Bxy = 0, Byx = 0.32, add.timestep = FALSE, plot = FALSE)
}
\arguments{
\item{n}{The number of time steps that will be generated.}

\item{xi}{The initial value for x}

\item{yi}{The initival value for y}

\item{Rx}{The value of the parameter Rx}

\item{Ry}{The value of the parameter Ry}

\item{Bxy}{Parameter controlling the sensitivity of x to changes in y.}

\item{Byx}{Parameter controlling the sensitivity of y to changes in x.}

\item{n}{Number of points in the time series.}
}
\value{
A dataframe containing columns time series for x and y
    (and optionally "t", a timestep column).

Defaults values are as in the original paper for figure 2A. Figure 2B in the original paper
uses 1000 different parameterizations with R and B chosen on the intervals
[3.6, 4.0] and [0, 0.4], respectively.
}
\description{
Source: Sugihara, G., May, R., Ye, H., Hsieh, C. H., Deyle, E., Fogarty, M., & Munch, S. (2012).
Detecting causality in complex ecosystems. Science, 338(6106), 496-500.
}
