% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic-map-twospecies-unidirectional-lagged.R
\name{logistic_map_twospecies_unidir_lagged}
\alias{logistic_map_twospecies_unidir_lagged}
\title{Time series for a two-species model system. Unidirectional forcing from x introduces synchrony in y.}
\usage{
logistic_map_twospecies_unidir_lagged(n = 3000, xi = sample(seq(0.01, 0.99,
  0.001), 1), yi = sample(seq(0.01, 0.99, 0.001), 1), Rx = sample(seq(3.57,
  3.82, 0.001), size = 1), Rxy = sample(seq(0.7, 0.9, 0.001), size = 1),
  Ry = sample(seq(3.57, 3.82, 0.001), size = 1), time.delay = 0,
  add.timestep = FALSE, plot = FALSE)
}
\arguments{
\item{n}{The number of time steps that will be generated.}

\item{xi}{Initial value of time series x.}

\item{yi}{Initial value of time series y.}

\item{Rx}{Parameter controlling the influence of x at timestep i on itself at timestep i+1}

\item{Rxy}{Parameter controlling the influence of x at timestep i on y at timestep i+delay}

\item{Ry}{Parameter controlling the influence of y at timestep i on itself at timestep i+1}

\item{add.timestep}{Whether or not to add a timestep column to the returned data frame.
Defaults to FALSE.}

\item{plot}{Whether to plot the dataset as a scatterplot matrix. Defaults to FALSE}

\item{delay}{The delay of the effect of x on y}
}
\value{
A dataframe containing columns time series for x and y
    (and optionally "t", a timestep column).
}
\description{
Source: Ye, H. et al. (2015). Distinguishing time-delayed causal interactions
using convergent cross mapping. Sci. Rep. 5, 14750; doi: 10.1038/srep14750 (2015).
Default values are as in the original paper.
}
