% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinkerbell-map.R
\name{tinkerbell_map}
\alias{tinkerbell_map}
\title{The tinkerbell map. From the Wikipedia article on chaotic maps.}
\usage{
tinkerbell_map(n = 1000, xi = -0.74, yi = -0.56, a = 0.9, b = -0.6,
  c = 2, d = 0.5, add.timestep = FALSE, plot = FALSE)
}
\arguments{
\item{n}{Number of points in the time series.}

\item{xi}{The initial value for x}

\item{yi}{The initival value for y}

\item{a}{The value of the parameter a}

\item{b}{The value of the parameter b}

\item{c}{The value of the parameter c}

\item{d}{The value of the parameter d}

\item{add.timestep}{Add a column indicating the time step?}

\item{plot}{Plot the time series?}
}
\value{
A n-by-2 data frame containing the values of the x and y components.
}
\description{
The tinkerbell map. From the Wikipedia article on chaotic maps.
}
